arm64: dts: qcom: sdm630: nile: Add support for Synaptics touchscreen
This adds support for Synaptics RMI4 touchscreen on
Sony Nile devices. It requires the usage of i2c-gpio
as just adding it to the i2c1 node does not seem to work.
Disable it on Ganges as it uses a Novatek touchscreen
instead.

This issue is known to happen on different devices, like
Oneplus 6/6T.

Signed-off-by: Konrad Dybcio <konradybcio@gmail.com>