DEVELOPER PREVIEW: Electrical spec V3, Adreno GPUs working
This is a developer preview -- I have no time to currently clean
up this mess because reasons.

This commit has been made public in order for other developers to
be able to continue their upstreaming work until I will be able to
cleanup this big piece of contribution and format proper commits.

Please, use it wisely!