interconnect: qcom: sdm660: Add SDM660 interconnect driver
Implement the interconnect on SDM660-class SoCs (630/636/660):.
These chips are managing the bus QoS in a "hybrid" fashion:
some of the paths are managed by (and through) the RPM uC, while
some are "AP Owned", meaning that the AP shall do direct writes
to the appropriate QoS registers for the specific paths and ports.

This driver also contains "a hint" of the QoS Limiter Mode: this
one is being used in other SDM/MSM SoCs while in SDM630/636/660
it appears to be unused, but the definition of this mode is added
anyway in this driver for the sake of complete information.

Signed-off-by: AngeloGioacchino Del Regno <kholk11@gmail.com>