arm64: dts: sdm630: Fix Adreno and Adreno SMMU clock tree
The clock tree for adreno was wrong: the GFX3D clock source has to be
the last one to bring up when enabling them and the first one to tear
down when disabling them (you really can't keep it on while the bus
and interface clocks are down!!).

Also, assign the right clocks to the Adreno SMMU node.
This makes possible to finally stop using clock hacks in the Adreno code.