[DONOTUPSTREAM] cpufreq: qcom-cpufreq-hw: Implement support for OSM programming
THIS IS A PROTOTYPE IMPLEMENTATION AND DOES WORK ONLY WITH SPECIFIC
PARAMETERS. DO NOT USE THIS DRIVER.

Some SoCs are usually shipped with firmware that does *not* program
the OSM controller in the bootloader, before jumping to the OS, nor
in the TZ environment.

In this case, the OSM programming should be performed by the operating
system: implement the procedure in the qcom-cpufreq-hw driver, which
is used to control the same hardware on devices that were already set
up fully from the bootloader.
This is "just an addition", meaning that the state machine logic of
this driver has not changed at all: infact, it's made to control the
exact same controller, just.. pre-programmed.